(Refact) seperate api calling logic from Competition.tsx
